+++ back/package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
-    "back": "nodemon index.js",
+    "start": "./node_modules/.bin/cross-env NODE_PATH=src NODE_ENV=development ./node_modules/.bin/nodemon src/index.js",
     "test": "echo \"ERROR: no test specified\" && exit 1"
   },
   "author": "junheeyeap",
@@ -24,6 +23,7 @@
     "passport-local": "^1.0.0"
   },
   "devDependencies": {
+    "cross-env": "^7.0.3",
     "nodemon": "^2.0.15"
   }
 }
