+++ back/src/routes/v1/index.js
@@ -1,0 +1,6 @@
+const express = require('express')
+
+const router = express.Router()
+router.use("/users", require("routes/v1/user"))
+
+module.exports = router
